Add dock for pack. Move JSONWarehouse in its own file
